Crate airbender_guest

Guest-side APIs for reading inputs and committing outputs.

Re-exports

pub use commit::commit;

pub use commit::exit_error;

pub use cycle::marker as cycle_marker;

pub use cycle::record_cycles;

pub use input::read;

pub use input::read_with;

pub use input::GuestError;

pub use transport::CsrTransport;

pub use transport::MockTransport;

pub use transport::Transport;

Modules

commit

Commit helpers for mapping values into output registers.

cycle

Development-only helpers for placing cycle markers in guest code.

input

Guest input helpers backed by the Airbender codec.

transport

Word-based transports for guest communication.

Traits

Commit

Values that can be committed to the public output registers (x10..x17).